Added login functionality to login.php, still some TODOs, #11
the passwords of the demo accounts in the database need to be
hashed first then the password_verify() function should be
placed in the if condition after that. this is the final step
the login process. Moreover, the index page should be configured
for the user based on the user type and the information that
should be showed to him. the functionality of index.php is
under issue #14.